UHF-13337: Test fixes
